requirements.txt
Faker
pytest
requests
selenium
webdriver-manager
allure-pytest
